+++ src/main/java/langley/storage/Storage.java
@@ -1,0 +1,42 @@
+package langley.storage;
+
+import langley.tasks.Task;
+
+import java.io.*;
+import java.util.ArrayList;
+
+public class Storage {
+    private String filePath;
+
+    public Storage(String filePath) {
+        this.filePath = filePath;
+    }
+
+    public void saveTasks(ArrayList<Task> tasks) {
+        try (BufferedWriter writer = new BufferedWriter(new FileWriter(filePath))) {
+            for (Task task : tasks) {
+                writer.write(task.toString());
+                writer.newLine();
+            }
+        } catch (IOException e) {
+            System.out.println("Error: Unable to save tasks to file.");
+        }
+    }
+
+    public void loadTasks(ArrayList<Task> tasks) {
+        File file = new File(filePath);
+        if (file.exists()) {
+            try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
+                String line;
+                while ((line = reader.readLine()) != null) {
+                    Task task = Task.fromString(line);
+                    if (task != null) {
+                        tasks.add(task);
+                    }
+                }
+            } catch (IOException e) {
+                System.out.println("Error: Unable to load tasks from file.");
+            }
+        }
+    }
+}
