+++ src/main/java/langley/tasklist/TaskList.java
@@ -1,0 +1,39 @@
+package langley.tasklist;
+
+import langley.tasks.Task;
+
+import java.util.ArrayList;
+
+public class TaskList {
+    private ArrayList<Task> tasks = new ArrayList<>();
+
+    public ArrayList<Task> getTasks() {
+        return tasks;
+    }
+
+    public Task getTask(int index) {
+        return tasks.get(index);
+    }
+
+    public void addTask(Task task) {
+        tasks.add(task);
+    }
+
+    public void deleteTask(int index) {
+        tasks.remove(index);
+    }
+
+    public void markTask(int index) {
+        tasks.get(index).mark();
+    }
+
+    public void unmarkTask(int index) {
+        tasks.get(index).unmark();
+    }
+
+    public void listTasks() {
+        for (int i = 0; i < tasks.size(); i++) {
+            System.out.println((i + 1) + ". " + tasks.get(i));
+        }
+    }
+}
