+++ src/main/java/langley/tasks/ToDo.java
@@ -1,0 +1,12 @@
+package langley.tasks;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String getType() {
+        return "T";
+    }
+}
